Updated Scanner and ScannerManager classes, added delete button in Scanner.xaml
Updated the `Scanner` class to include a new `Manager` property. Modified the `ScannerManager` class to include a new constructor, updated the `Add` method, and renamed the `RemoveScanner` method to `Remove`. Added a new delete button in `Scanner.xaml` with a command binding to `Manager.RemoveCommand` and a tooltip text "Delete This Camera?".