Refactor UI and optimize data binding
This commit introduces a series of enhancements aimed at improving the user interface's responsiveness and the efficiency of data binding across the application. Key changes include the simplification of the `Convert` and `ConvertBack` methods in the `ObjectIsNull` class through the use of expression-bodied members, and the modification of the `V52ImageTotalPixelDeviation` property in the `ImageEntry` class to be observable with a private setter, adhering to naming conventions. The inclusion of the `CommunityToolkit.Mvvm.ComponentModel` namespace in `IImageResultEntry.cs` supports MVVM components, while modifications to the `ImageResultEntry` and `ImageResultEntry_V275` classes streamline property change notifications and improve performance by initializing printer variables only when necessary. Typos have been corrected, and XAML files have been updated to enhance data binding, layout adjustments, and styling for a more intuitive and user-friendly UI. Additionally, adjustments in `ImageResultEntry_V5.xaml` such as DockPanel MaxHeight changes, data binding mode updates, and simplifications in icon sizes contribute to a more adaptable and visually appealing interface. The implementation of a new `ShortenString` converter and adjustments in namespace and converter usage further refine the application's visual presentation and data handling.