Add btnUnselect button and update event handlers
Added a new Button with the Click event handler btnUnselect to both Verifier.xaml and Scanner.xaml. This button is styled using MahApps.Styles.Button.Hamburger and is only visible when the IsDocked property of the parent UserControl is True. Modified the drwSettings_DrawerClosing method in Verifier.xaml.cs to use an expression-bodied member and execute the SaveCommand of the Manager property from the Verifier ViewModel. Added a new btnUnselect method in both Verifier.xaml.cs and Scanner.xaml.cs to handle the Click event of the newly added button. Currently, these methods are empty.