`Updated GS1 check, fixed Decode string, and modified build command`
Updated `Report.cs` to improve GS1 data check by using `GetMultipleKeyValuePair` function instead of checking the second element of `sym`. Moved the loop processing each string in `res` inside the `if (res != null)` block.

In `SectorDifferences.cs`, corrected the string used to get values for the "Decode" grade from "Decode" to "Decode,".

In `LabelVal.csproj`, updated the command executed after the build and the path to the `gs1encoders.dll` file.

Added a large amount of HTML and JavaScript code to `UpgradeLog.htm` for a report on project migration, including styles, functions for user interaction, and report content.